+++ courses/views_test.py
@@ -5,11 +5,9 @@ Tests for course views
 import operator as op
 
 import pytest
-import random
 import reversion
 from django.db.models import Count, Q
 from django.urls import reverse
-from pytest_django.fixtures import django_assert_max_num_queries
 from requests import ConnectionError as RequestsConnectionError
 from requests import HTTPError
 from rest_framework import status
@@ -21,14 +19,10 @@ from courses.factories import (
     CourseFactory,
     CourseRunEnrollmentFactory,
     CourseRunFactory,
-    ProgramFactory,
-    ProgramRequirementFactory,
 )
 from courses.models import (
     CourseRun,
     ProgramEnrollment,
-    ProgramRequirementNodeType,
-    ProgramRequirement,
 )
 from courses.serializers import (
     CourseRunEnrollmentSerializer,
@@ -88,93 +82,6 @@ def _num_queries_from_course(course):
     return (9 * num_programs) + (num_course_runs * 6) + 22
 
 
-@pytest.fixture()
-def programs():
-    """Fixture for a set of Programs in the database"""
-    return ProgramFactory.create_batch(3)
-
-
-@pytest.fixture()
-def courses():
-    """Fixture for a set of Courses in the database"""
-    return CourseFactory.create_batch(3)
-
-
-@pytest.fixture()
-def course_runs():
-    """Fixture for a set of CourseRuns in the database"""
-    return CourseRunFactory.create_batch(3)
-
-
-def populate_course_catalog_data(num_courses, num_programs):
-    """
-    Current production data is around 85 courses and 150 course runs. I opted to create 3 of each to allow
-    the best course run logic to play out as well as to push the endpoint a little harder in testing.
-
-    There are currently 3 programs in production, so I went with 15, again, to get ready for more data. To allow
-    things to be somewhat random, I grab courses for them at random (one required and one elective) which also allows
-    for courses to be in more than one program. If we need/want more specific test cases, we can add them, but this
-    is a more robust data set than production is presently.
-
-    Returns 3 separate lists to simulate what the tests received prior.
-
-    Args:
-        num_courses(int): number of courses to generate.
-        num_programs(int): number of programs to generate.
-    """
-    programs = []
-    courses = []
-    course_runs = []
-    for n in range(num_courses):
-        course, course_runs_for_course = _create_course(n)
-        courses.append(course)
-        course_runs.append(course_runs_for_course)
-    for n in range(num_programs):
-        program = _create_program(courses)
-        programs.append(program)
-    return courses, programs, course_runs
-
-
-def _create_course(n):
-    test_course = CourseFactory.create(title=f"Test Course {n}")
-    cr1 = CourseRunFactory.create(course=test_course, past_start=True)
-    cr2 = CourseRunFactory.create(course=test_course, in_progress=True)
-    cr3 = CourseRunFactory.create(course=test_course, in_future=True)
-    return test_course, [cr1, cr2, cr3]
-
-
-def _create_program(courses):
-    program = ProgramFactory.create()
-    ProgramRequirementFactory.add_root(program)
-    root_node = program.requirements_root
-    required_courses_node = root_node.add_child(
-        node_type=ProgramRequirementNodeType.OPERATOR,
-        operator=ProgramRequirement.Operator.ALL_OF,
-        title="Required Courses",
-    )
-    elective_courses_node = root_node.add_child(
-        node_type=ProgramRequirementNodeType.OPERATOR,
-        operator=ProgramRequirement.Operator.MIN_NUMBER_OF,
-        operator_value=2,
-        title="Elective Courses",
-        elective_flag=True,
-    )
-    if len(courses) > 3:
-        for c in random.sample(courses, 3):
-            required_courses_node.add_child(
-                node_type=ProgramRequirementNodeType.COURSE, course=c
-            )
-        for c in random.sample(courses, 3):
-            elective_courses_node.add_child(
-                node_type=ProgramRequirementNodeType.COURSE, course=c
-            )
-    else:
-        required_courses_node.add_child(
-            node_type=ProgramRequirementNodeType.COURSE, course=courses[0]
-        )
-    return program
-
-
 def test_get_programs(user_drf_client, programs):
     """Test the view that handles requests for all Programs"""
     resp = user_drf_client.get(reverse("programs_api-list"))
@@ -222,17 +129,16 @@ def test_delete_program(user_drf_client, programs):
     assert resp.status_code == status.HTTP_405_METHOD_NOT_ALLOWED
 
 
-@pytest.mark.parametrize("num_courses", [100])
-@pytest.mark.parametrize("num_programs", [15])
+@pytest.mark.parametrize("course_catalog_course_count", [100], indirect=True)
+@pytest.mark.parametrize("course_catalog_program_count", [15], indirect=True)
 def test_get_courses(
     user_drf_client,
     mock_context,
     django_assert_max_num_queries,
-    num_courses,
-    num_programs,
+    course_catalog_data
 ):
     """Test the view that handles requests for all Courses"""
-    courses, _, _ = populate_course_catalog_data(num_courses, num_programs)
+    courses, _, _ = course_catalog_data
     courses_from_fixture = []
     num_queries = 0
     for course in courses:
@@ -254,20 +160,16 @@ def test_get_courses(
     assert_drf_json_equal(courses_data, courses_from_fixture, ignore_order=True)
 
 
-@pytest.mark.parametrize("num_courses", [1])
-@pytest.mark.parametrize("num_programs", [1])
+@pytest.mark.parametrize("course_catalog_course_count", [1], indirect=True)
+@pytest.mark.parametrize("course_catalog_program_count", [1], indirect=True)
 def test_get_course(
     user_drf_client,
-    num_courses,
-    num_programs,
+   course_catalog_data,
     mock_context,
     django_assert_max_num_queries,
 ):
     """Test the view that handles a request for single Course"""
-    courses, _, _ = populate_course_catalog_data(
-        num_courses,
-        num_programs,
-    )
+    courses, _, _ = course_catalog_data
     course = courses[0]
     num_queries = _num_queries_from_course(course)
     with django_assert_max_num_queries(num_queries) as context:
@@ -282,17 +184,16 @@ def test_get_course(
     assert_drf_json_equal(course_data, course_from_fixture, ignore_order=True)
 
 
-@pytest.mark.parametrize("num_courses", [1])
-@pytest.mark.parametrize("num_programs", [1])
+@pytest.mark.parametrize("course_catalog_course_count", [1], indirect=True)
+@pytest.mark.parametrize("course_catalog_program_count", [1], indirect=True)
 def test_create_course(
     user_drf_client,
-    num_courses,
-    num_programs,
+    course_catalog_data,
     mock_context,
     django_assert_max_num_queries,
 ):
     """Test the view that handles a request to create a Course"""
-    courses, _, _ = populate_course_catalog_data(num_courses, num_programs)
+    courses, _, _ = course_catalog_data
     course = courses[0]
     course_data = CourseWithCourseRunsSerializer(
         instance=course, context=mock_context
@@ -306,16 +207,13 @@ def test_create_course(
     assert resp.status_code == status.HTTP_405_METHOD_NOT_ALLOWED
 
 
-@pytest.mark.parametrize("num_courses", [1])
-@pytest.mark.parametrize("num_programs", [1])
+@pytest.mark.parametrize("course_catalog_course_count", [1], indirect=True)
+@pytest.mark.parametrize("course_catalog_program_count", [1], indirect=True)
 def test_patch_course(
-    user_drf_client, num_courses, num_programs, django_assert_max_num_queries
+    user_drf_client, course_catalog_data, django_assert_max_num_queries
 ):
     """Test the view that handles a request to patch a Course"""
-    courses, _, _ = populate_course_catalog_data(
-        num_courses,
-        num_programs,
-    )
+    courses, _, _ = course_catalog_data
     course = courses[0]
     request_url = reverse("courses_api-detail", kwargs={"pk": course.id})
     with django_assert_max_num_queries(1) as context:
@@ -324,16 +222,13 @@ def test_patch_course(
     assert resp.status_code == status.HTTP_405_METHOD_NOT_ALLOWED
 
 
-@pytest.mark.parametrize("num_courses", [1])
-@pytest.mark.parametrize("num_programs", [1])
+@pytest.mark.parametrize("course_catalog_course_count", [1], indirect=True)
+@pytest.mark.parametrize("course_catalog_program_count", [1], indirect=True)
 def test_delete_course(
-    user_drf_client, num_courses, num_programs, django_assert_max_num_queries
+    user_drf_client, course_catalog_data, django_assert_max_num_queries
 ):
     """Test the view that handles a request to delete a Course"""
-    courses, _, _ = populate_course_catalog_data(
-        num_courses,
-        num_programs,
-    )
+    courses, _, _ = course_catalog_data
     course = courses[0]
     with django_assert_max_num_queries(1) as context:
         resp = user_drf_client.delete(
